pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^3.2"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^4.6.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.8.6"
sentry-sdk = "^1.3.1"
drf-spectacular = "^0.26.0"
django-tinymce = "^3.6.1"
python-dateutil = ">=2.4"
metadatax = {url = "https://github.com/PAM-Standardization/metadatax/archive/refs/tags/v0.4.12.tar.gz"}
#metadatax = {path = "../metadatax", develop=true}
whitenoise = "^6.9.0"
xarray = "^2024.10.0"
netcdf4 = "^1.6.5"
soundfile = "^0.12.1"
Pillow = "^10.0.0"
graphene-django = "^3.2.3"
graphene-django-pagination = {url = "https://github.com/instruct-br/graphene-django-pagination/archive/master.zip"}
django-filter = ">=2"
graphene-django-optimizer = "^0.10.0"
django-graphene-auth = "^1.1.1"
django-seed = "^0.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
coverage = { extras = ["toml"], version = "^6.3" }
black = "^22.10.0"
pylint = "3.3.7"
pylint-django = "2.6.1"
django-debug-toolbar = "^4.0.0"
faker = "^19.3.1"
freezegun = "^1.2.2"
django-seed = "^0.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source = ["backend"]
branch = true
omit = ["backend/asgi.py", "backend/wsgi.py", "backend/settings/*", "backend/api/migrations/*", "**/__init__.py"]

[tool.coverage.report]
show_missing = true

[tool.pylint.messages_control]
disable = [
    "too-many-ancestors",
    "too-many-arguments",
    "too-many-positional-arguments",
    "too-many-branches",
    "too-many-function-args",
    "arguments-renamed",
    "arguments-differ",
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "too-few-public-methods",
    "redefined-builtin",
    "unused-argument",
    "no-member",
]


[tool.pylint.MASTER]
load-plugins = "pylint_django"
django-settings-module = "backend.settings"

[tool.pylint.MAIN]
ignore-paths = ["backend\\api\\migrations|backend/api/migrations", "backend\\osmosewebsite\\migrations|backend/osmosewebsite/migrations"]

[tool.pylint.FORMAT]
max-line-length = 120
good-names = ["pk"]

[tool.pylint.DESIGN]
max-locals = 20
